Struct cranelift_codegen_meta::cdsl::ast::Apply

pub(crate) struct Apply {
    pub inst: Rc<InstructionContent>,
    pub args: Vec<Expr>,
    pub value_types: Vec<ValueType>,
}

Apply an instruction to arguments.

An Apply AST expression is created by using function call syntax on instructions. This applies to both bound and unbound polymorphic instructions.

Fields
inst: Rc<InstructionContent>args: Vec<Expr>value_types: Vec<ValueType>
Implementations

impl Apply

pub fn new(target: InstSpec, args: Vec<Expr>) -> Self

fn to_comment_string(&self, var_pool: &VarPool) -> String

pub fn inst_predicate(&self, var_pool: &VarPool) -> InstructionPredicate

pub fn inst_predicate_with_ctrl_typevar(
&self,
var_pool: &VarPool
) -> InstructionPredicate

Same as inst_predicate(), but also check the controlling type variable.

pub fn rust_builder(
&self,
defined_vars: &[VarIndex],
var_pool: &VarPool
) -> String