Struct cranelift_codegen_meta::cdsl::ast::Var

pub(crate) struct Var {
    pub name: String,
    pub src_def: Option<DefIndex>,
    pub dst_def: Option<DefIndex>,
    type_var: Option<TypeVar>,
    is_original_type_var: bool,
}

A free variable.

When variables are used in XForms with source and destination patterns, they are classified as follows:

Input values: Uses in the source pattern with no preceding def. These may appear as inputs in the destination pattern too, but no new inputs can be introduced.

Output values: Variables that are defined in both the source and destination pattern. These values may have uses outside the source pattern, and the destination pattern must compute the same value.

Intermediate values: Values that are defined in the source pattern, but not in the destination pattern. These may have uses outside the source pattern, so the defining instruction can’t be deleted immediately.

Temporary values are defined only in the destination pattern.

Fields
name: Stringsrc_def: Option<DefIndex>

The Def defining this variable in a source pattern.

dst_def: Option<DefIndex>

The Def defining this variable in a destination pattern.

type_var: Option<TypeVar>

TypeVar representing the type of this variable.

is_original_type_var: bool

Is this the original type variable, or has it be redefined with set_typevar?

Implementations

impl Var

fn new(name: String) -> Self

pub fn is_input(&self) -> bool

Is this an input value to the src pattern?

pub fn is_output(&self) -> bool

Is this an output value, defined in both src and dst patterns?

pub fn is_intermediate(&self) -> bool

Is this an intermediate value, defined only in the src pattern?

pub fn is_temp(&self) -> bool

Is this a temp value, defined only in the dst pattern?

pub fn get_def(&self, position: PatternPosition) -> Option<DefIndex>

Get the def of this variable according to the position.

pub fn set_def(&mut self, position: PatternPosition, def: DefIndex)

pub fn get_or_create_typevar(&mut self) -> TypeVar

Get the type variable representing the type of this variable.

pub fn get_typevar(&self) -> Option<TypeVar>

pub fn set_typevar(&mut self, tv: TypeVar)

pub fn has_free_typevar(&self) -> bool

Check if this variable has a free type variable. If not, the type of this variable is computed from the type of another variable.

pub fn to_rust_code(&self) -> String

fn rust_type(&self) -> String