Module cranelift_codegen_meta::cdsl::formats

Structs

FormatField 🔒

An immediate field in an instruction format.

FormatStructure 🔒

A tuple serving as a key to deduplicate InstructionFormat.

InstructionFormat 🔒

Every instruction opcode has a corresponding instruction format which determines the number of operands and their kinds. Instruction formats are identified structurally, i.e., the format of an instruction is derived from the kinds of operands used in its declaration.

InstructionFormatBuilder 🔒