Module cranelift_codegen_meta::cdsl::instructions

Structs

InstructionBuilder 🔒

InstructionContent 🔒

InstructionGroupBuilder 🔒

PolymorphicInfo 🔒

Functions

is_ctrl_typevar_candidate 🔒

Verify that the use of TypeVars is consistent with ctrl_typevar as the controlling type variable.

verify_format 🔒

Checks that the input operands actually match the given format.

verify_polymorphic 🔒

Check if this instruction is polymorphic, and verify its use of type variables.

Type Aliases

AllInstructions 🔒

Instruction 🔒