Struct cranelift_codegen_meta::cdsl::instructions::InstructionPredicate

pub(crate) struct InstructionPredicate {
    node: Option<InstructionPredicateNode>,
}

Fields
node: Option<InstructionPredicateNode>
Implementations

impl InstructionPredicate

pub fn new() -> Self

pub fn unwrap(self) -> InstructionPredicateNode

pub fn new_typevar_check(
inst: &Rc<InstructionContent>,
type_var: &TypeVar,
value_type: &ValueType
) -> InstructionPredicateNode

pub fn new_ctrl_typevar_check(
value_type: &ValueType
) -> InstructionPredicateNode

pub fn new_is_field_equal(
format: &InstructionFormat,
field_name: &'static str,
imm_value: String
) -> InstructionPredicateNode

pub fn new_is_field_equal_ast(
format: &InstructionFormat,
field: &FormatField,
imm_value: String
) -> InstructionPredicateNode

Used only for the AST module, which directly passes in the format field.

pub fn new_is_signed_int(
format: &InstructionFormat,
field_name: &'static str,
width: usize,
scale: usize
) -> InstructionPredicateNode

pub fn new_is_unsigned_int(
format: &InstructionFormat,
field_name: &'static str,
width: usize,
scale: usize
) -> InstructionPredicateNode

pub fn new_is_zero_int(
format: &InstructionFormat,
field_name: &'static str
) -> InstructionPredicateNode

pub fn new_is_zero_32bit_float(
format: &InstructionFormat,
field_name: &'static str
) -> InstructionPredicateNode

pub fn new_is_zero_64bit_float(
format: &InstructionFormat,
field_name: &'static str
) -> InstructionPredicateNode

pub fn new_is_all_zeroes(
format: &InstructionFormat,
field_name: &'static str
) -> InstructionPredicateNode

pub fn new_is_all_ones(
format: &InstructionFormat,
field_name: &'static str
) -> InstructionPredicateNode

pub fn new_length_equals(
format: &InstructionFormat,
size: usize
) -> InstructionPredicateNode

pub fn new_is_colocated_func(
format: &InstructionFormat,
field_name: &'static str
) -> InstructionPredicateNode

pub fn new_is_colocated_data(formats: &Formats) -> InstructionPredicateNode

pub fn and(self, new_node: InstructionPredicateNode) -> Self

pub fn or(self, new_node: InstructionPredicateNode) -> Self

pub fn rust_predicate(&self, func_str: &str) -> Option<String>

pub fn type_predicate(&self, func_str: &str) -> Option<String>

Returns the type predicate if this is one, or None otherwise.

pub fn collect_leaves(&self) -> Vec<&InstructionPredicateNode>

Returns references to all the nodes that are leaves in the condition (i.e. by flattening AND/OR).