Module cranelift_codegen_meta::cdsl::settings

Structs

BoolSetting

BoolSettingIndex

Predicate

Preset

PresetIndex

ProtoPredicate

ProtoSetting
This is the information provided during building for a setting.

Setting

SettingGroup

SettingGroupBuilder

Enums

PredicateNode

PresetType

ProtoSpecificSetting
This is the basic information needed to track the specific parts of a setting when building them.

SpecificSetting

Type Definitions

SettingPredicateNumber