Function cranelift_codegen_meta::cdsl::type_inference::constrain_fixpoint

fn constrain_fixpoint(tv1: &TypeVar, tv2: &TypeVar)

Given typevars tv1 and tv2 (which could be derived from one another), constrain their typesets to be the same. When one is derived from the other, repeat the constrain process until a fixed point is reached.