Module cranelift_codegen_meta::cdsl::types

Cranelift ValueType hierarchy

Structs

DynamicVectorType 🔒

A concrete dynamic SIMD vector type.

LaneTypeIterator 🔒

An iterator for different lane types.

ReferenceType 🔒

Reference type is scalar type, but not lane type.

ReferenceTypeIterator 🔒

An iterator for different reference types.

VectorType 🔒

A concrete SIMD vector type.

Enums

LaneType 🔒

A concrete scalar type that can appear as a vector lane too.

ValueType 🔒

A concrete SSA value type.

Statics

RUST_NAME_PREFIX 🔒