Struct cranelift_codegen_meta::cdsl::typevar::TypeVar

pub(crate) struct TypeVar {
    content: Rc<RefCell<TypeVarContent>>,
}

Fields
content: Rc<RefCell<TypeVarContent>>
Implementations

impl TypeVar

pub fn new( name: impl Into<String>, doc: impl Into<String>, type_set: TypeSet ) -> Self

pub fn new_singleton(value_type: ValueType) -> Self

pub fn copy_from(other: &TypeVar, name: String) -> TypeVar

Get a fresh copy of self, named after name. Can only be called on non-derived typevars.

pub fn get_typeset(&self) -> TypeSet

Returns the typeset for this TV. If the TV is derived, computes it recursively from the derived function and the base’s typeset. Note this can’t be done non-lazily in the constructor, because the TypeSet of the base may change over time.

pub fn get_raw_typeset(&self) -> &TypeSet

Returns this typevar’s type set, assuming this type var has no parent.

pub fn singleton_type(&self) -> Option<ValueType>

If the associated typeset has a single type return it. Otherwise return None.

pub fn free_typevar(&self) -> Option<TypeVar>

Get the free type variable controlling this one.

pub fn derived(&self, derived_func: DerivedFunc) -> TypeVar

Create a type variable that is a function of another.

pub fn lane_of(&self) -> TypeVar

pub fn as_truthy(&self) -> TypeVar

pub fn half_width(&self) -> TypeVar

pub fn double_width(&self) -> TypeVar

pub fn split_lanes(&self) -> TypeVar

pub fn merge_lanes(&self) -> TypeVar

pub fn dynamic_to_vector(&self) -> TypeVar

pub fn narrower(&self) -> TypeVar

Make a new TypeVar that includes all types narrower than self.

pub fn wider(&self) -> TypeVar

Make a new TypeVar that includes all types wider than self.