Type Definition cranelift_codegen_meta::cdsl::typevar::RangeBound

type RangeBound = u16;

A set of types.

We don’t allow arbitrary subsets of types, but use a parametrized approach instead.

Objects of this class can be used as dictionary keys.

Parametrized type sets are specified in terms of ranges:

The permitted range of vector lanes, where 1 indicates a scalar type.

The permitted range of integer types.

The permitted range of floating point types, and

The permitted range of boolean types.

The ranges are inclusive from smallest bit-width to largest bit-width.

Finally, a type set can contain special types (derived from SpecialType) which can’t appear as lane types.