Module cranelift_codegen_meta::cdsl::xform

Structs

Transform
An instruction transformation consists of a source and destination pattern.

TransformGroup
A group of related transformations.

TransformGroupBuilder

TransformGroupIndex

TransformGroups

Functions

rewrite_def_list

rewrite_defined_vars
Given a list of symbols defined in a Def, rewrite them to local symbols. Yield the new locals.

rewrite_expr
Find all uses of variables in expr and replace them with our own local symbols.

var_index
Inserts, if not present, a name in the symbol_table. Then returns its index in the variable pool var_pool. If the variable was not present in the symbol table, then add it to the list of defined_vars.

Type Definitions

SymbolTable