Function cranelift_codegen_meta::cdsl::xform::var_index

fn var_index(
    name: &str, 
    symbol_table: &mut HashMap<String, VarIndex>, 
    defined_vars: &mut Vec<VarIndex>, 
    var_pool: &mut VarPool
) -> VarIndex

Inserts, if not present, a name in the symbol_table. Then returns its index in the variable pool var_pool. If the variable was not present in the symbol table, then add it to the list of defined_vars.