Function cranelift_codegen_meta::gen_binemit::unwrap_values

fn unwrap_values(
    args: &[OperandConstraint], 
    prefix: &str, 
    values_slice: &str, 
    fmt: &mut Formatter
) -> String

Emit code that unwraps values living in registers or stack slots.

:param args: Input or output constraints. :param prefix: Prefix to be used for the generated local variables. :param values: Name of slice containing the values to be unwrapped. :returns: Comma separated list of the generated variables