Function cranelift_codegen_meta::gen_encodings::emit_operand_constraints

fn emit_operand_constraints(
    registers: &IsaRegs, 
    recipe: &EncodingRecipe, 
    constraints: &[OperandConstraint], 
    field_name: &'static str, 
    tied_operands: &HashMap<usize, usize>, 
    fixed_registers: &HashSet<Register>, 
    fmt: &mut Formatter
)

Emit a struct field initializer for an array of operand constraints.

Note "fixed_registers" must refer to the other kind of operands (i.e. if we're operating on inputs, fixed_registers must contain the fixed output registers).