Module cranelift_codegen_meta::gen_inst

Generate instruction data (including opcodes, formats, builders, etc.).

Constants

TYPESET_LIMIT ğŸ”’

Functions

gen_arguments_method ğŸ”’

gen_bitset ğŸ”’

gen_bool_accessor ğŸ”’

gen_builder ğŸ”’

Generate a Builder trait with methods for all instructions.

gen_format_constructor ğŸ”’

Emit a method for creating and inserting an instruction format.

gen_formats ğŸ”’

Generate an instruction format enumeration.

gen_inst_builder ğŸ”’

Emit a method for generating the instruction inst.

gen_instruction_data ğŸ”’

Generate the InstructionData enum.

gen_instruction_data_impl ğŸ”’

Generate the boring parts of the InstructionData implementation.

gen_isle ğŸ”’

gen_isle_enum ğŸ”’

Generate an enum immediate in ISLE.

gen_member_inits ğŸ”’

Emit member initializers for an instruction format.

gen_opcodes ğŸ”’

gen_try_from ğŸ”’

gen_type_constraints ğŸ”’

Generate value type constraints for all instructions.

gen_typesets_table ğŸ”’

Generate the table of ValueTypeSets described by type_sets.

generate ğŸ”’

get_constraint ğŸ”’

Get the value type constraint for an SSA value operand, where ctrl_typevar is the controlling type variable.

iterable_to_string ğŸ”’

typeset_to_string ğŸ”’