Module cranelift_codegen_meta::gen_legalizer

Generate transformations to legalize instructions without encodings.

Functions

build_derived_expr

emit_dst_inst

emit_runtime_typecheck
Emit rust code for the given check.

gen_isa
Generate legalization functions for isa and add any shared TransformGroups encountered to shared_groups.

gen_transform
Emit code for transform, assuming that the opcode of transform’s root instruction has already been matched.

gen_transform_group

generate
Generate the legalizer files.

is_value_split
Determine if node represents one of the value splitting instructions: isplit or vsplit. These instructions are lowered specially by the legalize::split` module.

unwrap_inst
Given a Def node, emit code that extracts all the instruction fields from pos.func.dfg[iref].