Function cranelift_codegen_meta::gen_legalizer::unwrap_inst

fn unwrap_inst(transform: &Transform, fmt: &mut Formatter) -> bool

Given a Def node, emit code that extracts all the instruction fields from pos.func.dfg[iref].

Create local variables named after the Var instances in node.

Also create a local variable named predicate with the value of the evaluated instruction predicate, or true if the node has no predicate.