Module cranelift_codegen_meta::gen_types

Generate sources with type info.

This generates a types.rs file which is included in cranelift-codegen/ir/types.rs. The file provides constant definitions for the most commonly used types, including all of the scalar types.

This ensures that the metaprogram and the generated program see the same type numbering.

Functions

emit_dynamic_vectors 🔒

Emit definition for all dynamic vector types with bits total size.

emit_type 🔒

Emit a constant definition of a single value type.

emit_types 🔒

Emit types using the given formatter object.

emit_vectors 🔒

Emit definition for all vector types with bits total size.

generate 🔒

Generate the types file.