Struct cranelift_codegen_meta::isa::x86::encodings::PerCpuModeEncodings

pub(crate) struct PerCpuModeEncodings {
    pub enc32: Vec<Rc<EncodingContent>>,
    pub enc64: Vec<Rc<EncodingContent>>,
    pub recipes: PrimaryMap<EncodingRecipeNumber, EncodingRecipe>,
    recipes_by_name: HashMap<String, EncodingRecipeNumber>,
    pub inst_pred_reg: InstructionPredicateRegistry,
}

Fields
enc32: Vec<Rc<EncodingContent>>enc64: Vec<Rc<EncodingContent>>recipes: PrimaryMap<EncodingRecipeNumber, EncodingRecipe>recipes_by_name: HashMap<String, EncodingRecipeNumber>inst_pred_reg: InstructionPredicateRegistry
Implementations

impl PerCpuModeEncodings

fn new() -> Self

fn add_recipe(&mut self, recipe: EncodingRecipe) -> EncodingRecipeNumber

fn make_encoding<T>(
&mut self,
inst: InstSpec,
template: Template<'_>,
builder_closure: T
) -> Rc<EncodingContent> where
T: FnOnce(EncodingBuilder) -> EncodingBuilder,

fn enc32_func<T>(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>,
builder_closure: T
) where
T: FnOnce(EncodingBuilder) -> EncodingBuilder,

fn enc32(&mut self, inst: impl Into<InstSpec>, template: Template<'_>)

fn enc32_isap(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>,
isap: u8
)

fn enc32_instp(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>,
instp: InstructionPredicateNode
)

fn enc32_rec(
&mut self,
inst: impl Into<InstSpec>,
recipe: &EncodingRecipe,
bits: u16
)

fn enc64_func<T>(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>,
builder_closure: T
) where
T: FnOnce(EncodingBuilder) -> EncodingBuilder,

fn enc64(&mut self, inst: impl Into<InstSpec>, template: Template<'_>)

fn enc64_isap(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>,
isap: u8
)

fn enc64_instp(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>,
instp: InstructionPredicateNode
)

fn enc64_rec(
&mut self,
inst: impl Into<InstSpec>,
recipe: &EncodingRecipe,
bits: u16
)

fn enc_i32_i64(&mut self, inst: impl Into<InstSpec>, template: Template<'_>)

Adds I32/I64 encodings as appropriate for a typed instruction. The REX prefix is always inferred at runtime.

Add encodings for inst.i32 to X86_32. Add encodings for inst.i32 to X86_64 with optional, inferred REX. Add encodings for inst.i64 to X86_64 with a REX.W prefix.

fn enc_i32_i64_explicit_rex(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>
)

Adds I32/I64 encodings as appropriate for a typed instruction. All variants of REX prefix are explicitly emitted, not inferred.

Add encodings for inst.i32 to X86_32. Add encodings for inst.i32 to X86_64 with and without REX. Add encodings for inst.i64 to X86_64 with and without REX.

fn enc_b32_b64(&mut self, inst: impl Into<InstSpec>, template: Template<'_>)

Adds B32/B64 encodings as appropriate for a typed instruction. The REX prefix is always inferred at runtime.

Adds encoding for inst.b32 to X86_32. Adds encoding for inst.b32 to X86_64 with optional, inferred REX. Adds encoding for inst.b64 to X86_64 with a REX.W prefix.

fn enc_i32_i64_rex_only(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>
)

Add encodings for inst.i32 to X86_32. Add encodings for inst.i32 to X86_64 with a REX prefix. Add encodings for inst.i64 to X86_64 with a REX.W prefix.

fn enc_i32_i64_instp(
&mut self,
inst: &Rc<InstructionContent>,
template: Template<'_>,
instp: InstructionPredicateNode
)

Add encodings for inst.i32 to X86_32. Add encodings for inst.i32 to X86_64 with and without REX. Add encodings for inst.i64 to X86_64 with a REX.W prefix.

fn enc_r32_r64_instp(
&mut self,
inst: &Rc<InstructionContent>,
template: Template<'_>,
instp: InstructionPredicateNode
)

Add encodings for inst.r32 to X86_32. Add encodings for inst.r32 to X86_64 with and without REX. Add encodings for inst.r64 to X86_64 with a REX.W prefix.

fn enc_r32_r64_rex_only(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>
)

Add encodings for inst.r32 to X86_32. Add encodings for inst.r64 to X86_64 with a REX.W prefix.

fn enc_r32_r64_ld_st(
&mut self,
inst: &Rc<InstructionContent>,
w_bit: bool,
template: Template<'_>
)

fn enc_x86_64(
&mut self,
inst: impl Into<InstSpec> + Clone,
template: Template<'_>
)

Add encodings for inst to X86_64 with and without a REX prefix.

fn enc_x86_64_instp(
&mut self,
inst: impl Clone + Into<InstSpec>,
template: Template<'_>,
instp: InstructionPredicateNode
)

Add encodings for inst to X86_64 with and without a REX prefix.

fn enc_x86_64_isap(
&mut self,
inst: impl Clone + Into<InstSpec>,
template: Template<'_>,
isap: u8
)

fn enc_both(
&mut self,
inst: impl Clone + Into<InstSpec>,
template: Template<'_>
)

Add all three encodings for inst:

X86_32

X86_64 with and without the REX prefix.

fn enc_both_isap(
&mut self,
inst: impl Clone + Into<InstSpec>,
template: Template<'_>,
isap: u8
)

fn enc_both_instp(
&mut self,
inst: impl Clone + Into<InstSpec>,
template: Template<'_>,
instp: InstructionPredicateNode
)

fn enc_both_inferred(
&mut self,
inst: impl Clone + Into<InstSpec>,
template: Template<'_>
)

Add two encodings for inst:

X86_32, no REX prefix, since this is not valid in 32-bit mode.

X86_64, dynamically infer the REX prefix.

fn enc_both_inferred_maybe_isap(
&mut self,
inst: impl Clone + Into<InstSpec>,
template: Template<'_>,
isap: Option<u8>
)

fn enc_both_rex_only(
&mut self,
inst: impl Clone + Into<InstSpec>,
template: Template<'_>
)

Add two encodings for inst:

X86_32

X86_64 with the REX prefix.

fn enc_i32_i64_ld_st(
&mut self,
inst: &Rc<InstructionContent>,
w_bit: bool,
template: Template<'_>
)

Add encodings for inst.i32 to X86_32. Add encodings for inst.i32 to X86_64 with and without REX. Add encodings for inst.i64 to X86_64 with a REX prefix, using the w_bit argument to determine whether or not to set the REX.W bit.

fn enc_32_64_rec(
&mut self,
inst: impl Clone + Into<InstSpec>,
recipe: &EncodingRecipe,
bits: u16
)

Add the same encoding/recipe pairing to both X86_32 and X86_64

fn enc_32_64_func<T>(
&mut self,
inst: impl Clone + Into<InstSpec>,
template: Template<'_>,
builder_closure: T
) where
T: FnOnce(EncodingBuilder) -> EncodingBuilder,

Add the same encoding to both X86_32 and X86_64; assumes configuration (e.g. REX, operand binding) has already happened

fn enc_32_64_maybe_isap(
&mut self,
inst: impl Clone + Into<InstSpec>,
template: Template<'_>,
isap: Option<u8>
)

Add the same encoding to both X86_32 and X86_64; assumes configuration (e.g. REX, operand binding) has already happened.

fn enc32_maybe_isap(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>,
isap: Option<u8>
)

fn enc64_maybe_isap(
&mut self,
inst: impl Into<InstSpec>,
template: Template<'_>,
isap: Option<u8>
)