Struct cranelift_codegen_meta::shared::entities::EntityRefs

pub(crate) struct EntityRefs {
    pub(crate) block_call: OperandKind,
    pub(crate) label: OperandKind,
    pub(crate) block_then: OperandKind,
    pub(crate) block_else: OperandKind,
    pub(crate) stack_slot: OperandKind,
    pub(crate) dynamic_stack_slot: OperandKind,
    pub(crate) global_value: OperandKind,
    pub(crate) sig_ref: OperandKind,
    pub(crate) func_ref: OperandKind,
    pub(crate) jump_table: OperandKind,
    pub(crate) table: OperandKind,
    pub(crate) varargs: OperandKind,
}

Fields
block_call: OperandKind

A reference to a basic block in the same function, with its arguments provided. This is primarily used in control flow instructions.

label: OperandKind

A reference to a basic block in the same function. This is primarily used in control flow instructions.

block_then: OperandKind

A reference to a basic block in the same function, with its arguments provided. This is primarily used in control flow instructions.

block_else: OperandKind

A reference to a basic block in the same function, with its arguments provided. This is primarily used in control flow instructions.

stack_slot: OperandKind

A reference to a stack slot declared in the function preamble.

dynamic_stack_slot: OperandKind

A reference to a dynamic_stack slot declared in the function preamble.

global_value: OperandKind

A reference to a global value.

sig_ref: OperandKind

A reference to a function signature declared in the function preamble. This is used to provide the call signature in a call_indirect instruction.

func_ref: OperandKind

A reference to an external function declared in the function preamble. This is used to provide the callee and signature in a call instruction.

jump_table: OperandKind

A reference to a jump table declared in the function preamble.

table: OperandKind

A reference to a table declared in the function preamble.

varargs: OperandKind

A variable-sized list of value operands. Use for Block and function call arguments.

Implementations

impl EntityRefs

pub fn new() -> Self