Struct cranelift_codegen_meta::shared::Definitions

pub(crate) struct Definitions {
    pub settings: SettingGroup,
    pub all_instructions: PrimaryMap<OpcodeNumber, Rc<InstructionContent>>,
    pub instructions: InstructionGroup,
    pub imm: Immediates,
    pub formats: Formats,
    pub transform_groups: TransformGroups,
    pub entities: EntityRefs,
}

Fields
settings: SettingGroupall_instructions: PrimaryMap<OpcodeNumber, Rc<InstructionContent>>instructions: InstructionGroupimm: Immediatesformats: Formatstransform_groups: TransformGroupsentities: EntityRefs
Implementations

impl Definitions

pub fn verify_instruction_formats(&self) -> Vec<&InstructionFormat>

Verifies certain properties of formats.

Formats must be uniquely named: if two formats have the same name, they must refer to the same data. Otherwise, two format variants in the codegen crate would have the same name.

Formats must be structurally different from each other. Otherwise, this would lead to code duplicate in the codegen crate.

Returns a list of all the instruction formats effectively used.