Module cranelift_codegen_meta::shared::types

This module predefines all the Cranelift scalar types.

Structs

BoolIterator ğŸ”’

This provides an iterator through all of the supported bool variants.

FlagIterator ğŸ”’

Iterator through the variants of the Flag enum.

FloatIterator ğŸ”’

Iterator through the variants of the Float enum.

IntIterator ğŸ”’

This provides an iterator through all of the supported int variants.

ReferenceIterator ğŸ”’

This provides an iterator through all of the supported reference variants.

Enums

Bool ğŸ”’

Flag ğŸ”’

A type representing CPU flags.

Float ğŸ”’

Int ğŸ”’

Reference ğŸ”’