Module cranelift_codegen_meta::srcgen

Source code generator.

The srcgen module contains generic helper routines and classes for generating source code.

Macros

fmtln 🔒

A macro that simplifies the usage of the Formatter by allowing format strings.

Structs

Formatter 🔒

Match 🔒

Match formatting class.

Statics

SHIFTWIDTH 🔒

Functions

_indent 🔒

Compute the indentation of s, or None of an empty line.

parse_multiline 🔒

Given a multi-line string, split it into a sequence of lines after stripping a common indentation. This is useful for strings defined with doc strings.