Struct cranelift_codegen_meta::srcgen::Formatter

pub(crate) struct Formatter {
    indent: usize,
    lines: Vec<String>,
}

Fields
indent: usizelines: Vec<String>
Implementations

impl Formatter

pub fn new() -> Self

Source code formatter class. Used to collect source code to be written to a file, and keep track of indentation.

pub fn indent_push(&mut self)

Increase current indentation level by one.

pub fn indent_pop(&mut self)

Decrease indentation by one level.

pub fn indent<T, F: FnOnce(&mut Formatter) -> T>(&mut self, f: F) -> T

fn get_indent(&self) -> String

Get the current whitespace indentation in the form of a String.

fn get_outdent(&mut self) -> String

Get a string containing whitespace outdented one level. Used for lines of code that are inside a single indented block.

pub fn line(&mut self, contents: impl AsRef<str>)

Add an indented line.

pub fn empty_line(&mut self)

Pushes an empty line.

pub fn outdented_line(&mut self, s: &str)

Emit a line outdented one level.

pub fn update_file(
&self,
filename: impl AsRef<str>,
directory: &str
) -> Result<(), Error>

Write self.lines to a file.

pub fn multi_line(&mut self, s: &str)

Add one or more lines after stripping common indentation.

pub fn comment(&mut self, s: impl AsRef<str>)

Add a comment line.

pub fn doc_comment(&mut self, contents: impl AsRef<str>)

Add a (multi-line) documentation comment.

pub fn add_match(&mut self, m: Match)

Add a match expression.