Crate cranelift_codegen_meta

This crate generates Rust sources for use by cranelift_codegen.

Modules

error

Error returned during meta code-generation.

isa

Define supported ISAs; includes ISA-specific instructions, encodings, registers, settings, etc.

Macros

predicate

A macro that converts boolean settings into predicates to look more natural.

preset

A macro that joins boolean settings into a list (e.g. preset!(feature_a && feature_b)).

Functions

generate

Generates all the Rust source files used in Cranelift from the meta-language.

isa_from_arch

Generate an ISA from an architecture string (e.g. “x86_64”).