Crate cranelift_codegen_meta

This crate generates Rust sources for use by cranelift_codegen.

Modules

cdsl
Cranelift DSL classes.

default_map
Trait for extending HashMap with get_or_default.

error
Error returned during meta code-generation.

gen_binemit
Generate binary emission code for each ISA.

gen_encodings
Generate sources for instruction encoding.

gen_inst
Generate instruction data (including opcodes, formats, builders, etc.).

gen_legalizer
Generate transformations to legalize instructions without encodings.

gen_registers
Generate the ISA-specific registers.

gen_settings
Generate the ISA-specific settings.

gen_types
Generate sources with type info.

isa
Define supported ISAs; includes ISA-specific instructions, encodings, registers, settings, etc.

shared
Shared definitions for the Cranelift intermediate language.

srcgen
Source code generator.

unique_table
An index-accessed table implementation that avoids duplicate entries.

Macros

predicate
A macro that converts boolean settings into predicates to look more natural.

preset
A macro that joins boolean settings into a list (e.g. preset!(feature_a && feature_b)).

Functions

generate
Generates all the Rust source files used in Cranelift from the meta-language.

isa_from_arch
Generate an ISA from an architecture string (e.g. "x86_64").