Module cranelift_codegen_meta::cdsl::ast

Structs

Apply

Apply an instruction to arguments.

Block

A definition which would lead to generate a block creation.

BlockPool

ConstPool

Contains constants created in the AST that must be inserted into the true ConstantPool when the legalizer code is generated. The constant data is named in the order it is inserted; inserting data using insert will avoid duplicates.

Def

An AST definition associates a set of variables with the values produced by an expression.

DefIndex

DefPool

DummyConstant

DummyDef

DummyVar

ExprBuilder

Var

A free variable.

VarIndex

VarPool

Enums

DummyExpr

Expr

Literal

PatternPosition

Functions

constant

var