Module cranelift_codegen_meta::cdsl::instructions

Structs

InstructionBuilder ğŸ”’

InstructionContent ğŸ”’

InstructionGroupBuilder ğŸ”’

PolymorphicInfo ğŸ”’

Functions

is_ctrl_typevar_candidate ğŸ”’

Verify that the use of TypeVars is consistent with ctrl_typevar as the controlling type variable.

verify_format ğŸ”’

Checks that the input operands actually match the given format.

verify_polymorphic ğŸ”’

Check if this instruction is polymorphic, and verify its use of type variables.

Type Definitions

AllInstructions ğŸ”’

Instruction ğŸ”’