Struct cranelift_codegen_meta::cdsl::instructions::InstructionBuilder

pub(crate) struct InstructionBuilder {

    name: String,
    doc: String,
    format: Rc<InstructionFormat>,
    operands_in: Option<Vec<Operand>>,
    operands_out: Option<Vec<Operand>>,
    is_terminator: bool,
    is_branch: bool,
    is_call: bool,
    is_return: bool,
    can_load: bool,
    can_store: bool,
    can_trap: bool,
    other_side_effects: bool,
    side_effects_idempotent: bool,

}

Fields
name: Stringdoc: Stringformat: Rc<InstructionFormat>operands_in: Option<Vec<Operand>>operands_out: Option<Vec<Operand>>is_terminator: boolis_branch: boolis_call: boolis_return: boolcan_load: boolcan_store: boolcan_trap: boolother_side_effects: boolside_effects_idempotent: bool
Implementations

impl InstructionBuilder

pub fn new<S: Into<String>>( name: S, doc: S, format: &Rc<InstructionFormat> ) -> Self

pub fn operands_in(self, operands: Vec<Operand>) -> Self

pub fn operands_out(self, operands: Vec<Operand>) -> Self

pub fn terminates_block(self) -> Self

Mark this instruction as a block terminator.

pub fn branches(self) -> Self

Mark this instruction as a branch instruction. This also implies that the instruction is a block terminator.

pub fn call(self) -> Self

Mark this instruction as a call instruction.

pub fn returns(self) -> Self

Mark this instruction as a return instruction. This also implies that the instruction is a block terminator.

pub fn can_load(self) -> Self

Mark this instruction as one that can load from memory.

pub fn can_store(self) -> Self

Mark this instruction as one that can store to memory.

pub fn can_trap(self) -> Self

Mark this instruction as possibly trapping.

pub fn other_side_effects(self) -> Self

Mark this instruction as one that has side-effects.

pub fn side_effects_idempotent(self) -> Self

Mark this instruction as one whose side-effects may be de-duplicated.

fn build(self) -> Rc<InstructionContent>