Module cranelift_codegen_meta::cdsl::types

Cranelift ValueType hierarchy

Structs

LaneTypeIterator

An iterator for different lane types.

ReferenceType

Reference type is scalar type, but not lane type.

ReferenceTypeIterator

An iterator for different reference types.

SpecialTypeIterator

VectorType

A concrete SIMD vector type.

Enums

LaneType

A concrete scalar type that can appear as a vector lane too.

SpecialType

A concrete scalar type that is neither a vector nor a lane type.

ValueType

A concrete SSA value type.

Statics

RUST_NAME_PREFIX