Enum cranelift_codegen_meta::cdsl::types::ValueType

pub(crate) enum ValueType {
    Lane(LaneType),
    Reference(ReferenceType),
    Special(SpecialType),
    Vector(VectorType),
}

A concrete SSA value type.

All SSA values have a type that is described by an instance of ValueType or one of its subclasses.

Variants

Lane(LaneType)

Reference(ReferenceType)

Special(SpecialType)

Vector(VectorType)

Implementations

impl ValueType

pub fn all_lane_types() -> LaneTypeIterator

Iterate through all of the lane types.

pub fn all_special_types() -> SpecialTypeIterator

Iterate through all of the special types (neither lanes nor vectors).

pub fn all_reference_types() -> ReferenceTypeIterator

pub fn doc(&self) -> String

Return a string containing the documentation comment for this type.

pub fn lane_bits(&self) -> u64

Return the number of bits in a lane.

pub fn lane_count(&self) -> u64

Return the number of lanes.

pub fn membytes(&self) -> u64

Find the number of bytes that this type occupies in memory.

pub fn number(&self) -> Option<u8>

Find the unique number associated with this type.

pub fn rust_name(&self) -> String

Return the name of this type for generated Rust source files.

pub fn _wider_or_equal(&self, rhs: &ValueType) -> bool

Return true iff:

self and other have equal number of lanes

each lane in self has at least as many bits as a lane in other

pub fn width(&self) -> u64

Return the total number of bits of an instance of this type.