Module cranelift_codegen_meta::cdsl::typevar

Structs

TypeSet

TypeSetBuilder

TypeVar

TypeVarContent

Type variables can be used in place of concrete types when defining instructions. This makes the instructions polymorphic.

TypeVarParent

Enums

DerivedFunc

Interval

Constants

MAX_BITS

MAX_FLOAT_BITS

MAX_LANES

Functions

legal_bool

range_to_set

Generates a set with all the powers of two included in the range.

set_narrower

set_wider_or_equal

Type Definitions

NumSet

Range

RangeBound

A set of types.