Module cranelift_codegen_meta::cdsl::typevar

Macros

num_set ğŸ”’

Structs

TypeSet ğŸ”’

TypeSetBuilder ğŸ”’

TypeVar ğŸ”’

TypeVarContent ğŸ”’

Type variables can be used in place of concrete types when defining instructions. This makes the instructions polymorphic.

TypeVarParent ğŸ”’

Enums

DerivedFunc ğŸ”’

Interval ğŸ”’

Constants

MAX_BITS ğŸ”’

MAX_FLOAT_BITS ğŸ”’

MAX_LANES ğŸ”’

Functions

legal_bool ğŸ”’

range_to_set ğŸ”’

Generates a set with all the powers of two included in the range.

Type Definitions

NumSet ğŸ”’

Range ğŸ”’

RangeBound ğŸ”’

A set of types.