Struct cranelift_codegen_meta::cdsl::typevar::TypeSet

pub(crate) struct TypeSet {
    pub lanes: BTreeSet<u16>,
    pub ints: BTreeSet<u16>,
    pub floats: BTreeSet<u16>,
    pub bools: BTreeSet<u16>,
    pub refs: BTreeSet<u16>,
    pub specials: Vec<SpecialType>,
}

Fields
lanes: BTreeSet<u16>ints: BTreeSet<u16>floats: BTreeSet<u16>bools: BTreeSet<u16>refs: BTreeSet<u16>specials: Vec<SpecialType>
Implementations

impl TypeSet

fn new(
lanes: BTreeSet<u16>,
ints: BTreeSet<u16>,
floats: BTreeSet<u16>,
bools: BTreeSet<u16>,
refs: BTreeSet<u16>,
specials: Vec<SpecialType>
) -> Self

pub fn size(&self) -> usize

Return the number of concrete types represented by this typeset.

fn image(&self, derived_func: DerivedFunc) -> TypeSet

Return the image of self across the derived function func.

fn lane_of(&self) -> TypeSet

Return a TypeSet describing the image of self across lane_of.

fn as_bool(&self) -> TypeSet

Return a TypeSet describing the image of self across as_bool.

fn half_width(&self) -> TypeSet

Return a TypeSet describing the image of self across halfwidth.

fn double_width(&self) -> TypeSet

Return a TypeSet describing the image of self across doublewidth.

fn half_vector(&self) -> TypeSet

Return a TypeSet describing the image of self across halfvector.

fn double_vector(&self) -> TypeSet

Return a TypeSet describing the image of self across doublevector.

fn concrete_types(&self) -> Vec<ValueType>

fn get_singleton(&self) -> ValueType

Return the singleton type represented by self. Can only call on typesets containing 1 type.

fn preimage(&self, func: DerivedFunc) -> TypeSet

Return the inverse image of self across the derived function func.

pub fn inplace_intersect_with(&mut self, other: &TypeSet)

pub fn is_subset(&self, other: &TypeSet) -> bool

pub fn is_wider_or_equal(&self, other: &TypeSet) -> bool

pub fn is_narrower(&self, other: &TypeSet) -> bool