Module cranelift_codegen_meta::gen_inst

Generate instruction data (including opcodes, formats, builders, etc.).

Constants

TYPESET_LIMIT

Functions

gen_arguments_method

gen_bitset

gen_bool_accessor

gen_builder
Generate a Builder trait with methods for all instructions.

gen_format_constructor
Emit a method for creating and inserting an instruction format.

gen_formats
Generate an instruction format enumeration.

gen_inst_builder
Emit a method for generating the instruction inst.

gen_instruction_data
Generate the InstructionData enum.

gen_instruction_data_impl
Generate the boring parts of the InstructionData implementation.

gen_member_inits
Emit member initializers for an instruction format.

gen_opcodes

gen_type_constraints
Generate value type constraints for all instructions.

gen_typesets_table
Generate the table of ValueTypeSets described by type_sets.

generate

get_constraint
Get the value type constraint for an SSA value operand, where ctrl_typevar is the controlling type variable.

iterable_to_string

typeset_to_string